# openapi.yaml
openapi: 3.1.0
info:
  title: Anamnesis HTTP API
  version: 1.0.0
  summary: Narrow clinician-in-the-loop workflow API for case intake, bounded evidence ingestion, packet drafting, review, and audit.
  description: |
    The Anamnesis API is a narrow workflow surface for organizing intake, source artifacts,
    bounded text and FHIR import seams, physician packet drafts, explicit clinician review,
    packet finalization, and append-only audit history.

    This API is not a diagnosis, treatment, prescription, or general-purpose FHIR server API.

    Authentication model:
    - application endpoints expect `Authorization: Bearer <API_KEY>` when `API_KEY` is configured;
    - application endpoints also accept `Authorization: Bearer <JWT>` when `JWT_SECRET` or `JWT_PUBLIC_KEY` is configured;
    - `JWT_SECRET` and `JWT_PUBLIC_KEY` are mutually exclusive JWT verifier modes;
    - when `API_KEY` and one JWT verifier mode are configured, either bearer mechanism may be used;
    - JWTs must carry a non-empty `sub`, must satisfy configured `iss`/`aud` constraints, and may additionally be required to match a configured JOSE `typ` value;
    - `JWT_PUBLIC_KEY` enables RS256 verification and requires an RSA public key with modulus length >= 2048 bits;
    - unauthenticated startup is allowed only when `ALLOW_INSECURE_DEV_AUTH=true` is explicitly set;
    - the insecure development override is rejected when `NODE_ENV=production`.

    Interoperability boundary:
...
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
  - jwtBearer: []
paths:
...
  /api/cases/{caseId}/document-ingestions:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Ingestion]
      summary: Ingest bounded text document
      operationId: ingestDocument
      description: |
        Normalize a bounded `text/plain` or `text/markdown` payload into a source artifact.
        Request body parsing is bounded at the HTTP layer and the domain stores only a bounded excerpt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentIngestionRequest'
      responses:
        '201':
          description: Document ingested.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentIngestionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /api/cases/{caseId}/fhir-imports:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [FHIR]
      summary: Import one supported inline FHIR resource
      operationId: importFhirResource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FhirImportRequest'
      responses:
        '201':
          description: FHIR resource imported.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FhirImportResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse400'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /api/cases/{caseId}/fhir-bundle-imports:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [FHIR]
      summary: Import supported FHIR bundle entries
      operationId: importFhirBundle
      description: |
        Import a `document` or `collection` bundle containing supported `Binary` and `DocumentReference`
        entries. `document` bundles must include a `Composition` first entry plus `identifier.system`,
        `identifier.value`, and `timestamp`. External attachment fetch remains opt-in and heavily bounded.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FhirBundleImportRequest'
      responses:
        '201':
          description: FHIR bundle imported.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FhirBundleImportResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse400'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /api/cases/{caseId}/physician-packets:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Packets]
      summary: Draft physician packet
      operationId: createPhysicianPacket
      description: |
        Draft a physician packet from current case state. Under JWT bearer auth, `requestedBy` is bound to the
        authenticated JWT `sub` and mismatched caller-supplied identities are rejected. At least one artifact must
        already exist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePacketRequest'
      responses:
        '201':
          description: Packet drafted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicianPacketCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
    get:
      tags: [Packets]
      summary: List physician packets
      operationId: listPhysicianPackets
      responses:
        '200':
          description: Packets listed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicianPacketListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /api/cases/{caseId}/physician-packets/{packetId}/reviews:
    parameters:
      - $ref: '#/components/parameters/CaseId'
      - $ref: '#/components/parameters/PacketId'
    post:
      tags: [Reviews]
      summary: Submit clinician review
      operationId: submitReview
      description: |
        Append a review entry. When JWT bearer auth is used, `reviewerName` is bound to the authenticated JWT `sub`
        and the principal must carry either the `reviewer` or `clinician` role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitReviewRequest'
      responses:
        '201':
          description: Review appended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
    get:
      tags: [Reviews]
      summary: Get review ledger
      description: |
        Finalize a clinician-approved packet that is not stale. When JWT bearer auth is used, `finalizedBy` is bound
        to the authenticated JWT `sub` and the principal must carry the `clinician` role.
      responses:
        '200':
          description: Review ledger returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /api/cases/{caseId}/physician-packets/{packetId}/finalize:
    parameters:
      - $ref: '#/components/parameters/CaseId'
      - $ref: '#/components/parameters/PacketId'
    post:
      tags: [Packets]
      summary: Finalize physician packet
      operationId: finalizePhysicianPacket
      description: Finalize a clinician-approved packet that is not stale.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizePacketRequest'
      responses:
        '200':
          description: Packet finalized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicianPacketCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
...
  /api/audit-chain/verify:
    get:
      tags: [Audit]
      summary: Verify audit hash chain for one case
      operationId: verifyAuditChainByCase
      description: |
        Recompute and verify the SHA-256 audit hash chain for all events returned by one `caseId`.
        This is a read-only tamper-evidence check and does not mutate stored events.
      parameters:
        - name: caseId
          in: query
          required: true
          description: Case identifier whose append-only audit chain should be verified.
          schema:
            type: string
      responses:
        '200':
          description: Audit chain verification result returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditChainVerificationResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse400'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /api/operations/summary:
    get:
      tags: [Operations]
      summary: Get operations summary
      operationId: getOperationsSummary
      responses:
        '200':
          description: Operations summary returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsSummaryResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque API key
    jwtBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token validated via HS256 when `JWT_SECRET` is configured or via RS256 when `JWT_PUBLIC_KEY` is configured. Runtime validation enforces signature, `exp`, `iat`, `nbf`, non-empty `sub`, configured `iss`/`aud`, optional JOSE `typ`, and an RSA-2048 minimum for RS256 verification keys.
  parameters:
    CaseId:
      name: caseId
      in: path
      required: true
      description: Case UUID.
      schema:
        type: string
        format: uuid
...
    PacketId:
      name: packetId
      in: path
      required: true
      description: Physician packet UUID.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequestError:
      description: Request validation or request-governance error.
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ValidationErrorResponse'
              - $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    UnauthorizedError:
      description: Missing or invalid bearer token.
      headers:
        WWW-Authenticate:
          description: Bearer authentication challenge.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Authenticated principal is not allowed to perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Workflow or domain conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse400:
      description: Domain or parsing error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequestsError:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    DateOnly:
      type: string
      pattern: '^\d{4}-\d{2}-\d{2}$'
      examples: ['2026-04-03']
...
    ArtifactType:
      type: string
      enum: [note, lab, summary, report, imaging-summary]
...
    DocumentContentType:
      type: string
      enum: [text/plain, text/markdown]
...
    ReviewAction:
      type: string
      enum: [approved, changes_requested, rejected]
...
    AnamnesisCase:
      type: object
      additionalProperties: false
      required:
        - caseId
        - workflowFamily
        - status
        - createdAt
        - updatedAt
        - intake
        - samples
        - artifacts
        - physicianPackets
      properties:
        caseId:
          type: string
          format: uuid
        patientLabel:
          type: string
        workflowFamily:
          $ref: '#/components/schemas/WorkflowFamily'
        status:
          $ref: '#/components/schemas/CaseStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        intake:
          $ref: '#/components/schemas/CaseIntake'
        samples:
          type: array
          items:
            $ref: '#/components/schemas/SampleRecord'
        studyContext:
          $ref: '#/components/schemas/StudyContextRecord'
        qcSummary:
          $ref: '#/components/schemas/QcSummaryRecord'
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/SourceArtifact'
        physicianPackets:
          type: array
          items:
            $ref: '#/components/schemas/PhysicianPacket'
    DocumentIngestionResult:
      type: object
      additionalProperties: false
      required: [contentType, normalizationProfile, normalizedCharacterCount, excerptCharacterCount, truncated]
      properties:
        contentType:
          $ref: '#/components/schemas/DocumentContentType'
        normalizationProfile:
          $ref: '#/components/schemas/DocumentNormalizationProfile'
        filename:
          type: string
        normalizedCharacterCount:
          type: integer
          minimum: 0
        excerptCharacterCount:
          type: integer
          minimum: 0
        truncated:
          type: boolean
...
    DocumentIngestionResponse:
      type: object
      additionalProperties: false
      required: [case, artifact, ingestion]
      properties:
        case:
          $ref: '#/components/schemas/AnamnesisCase'
        artifact:
          $ref: '#/components/schemas/SourceArtifact'
        ingestion:
          $ref: '#/components/schemas/DocumentIngestionResult'
    FhirImportResponse:
      type: object
      additionalProperties: false
      required: [case, artifact, ingestion, fhirImport]
      properties:
        case:
          $ref: '#/components/schemas/AnamnesisCase'
        artifact:
          $ref: '#/components/schemas/SourceArtifact'
        ingestion:
          $ref: '#/components/schemas/DocumentIngestionResult'
        fhirImport:
          $ref: '#/components/schemas/FhirImportResult'
    FhirBundleImportResponse:
      type: object
      additionalProperties: false
      required: [case, artifacts, ingestions, bundleImport]
      properties:
        case:
          $ref: '#/components/schemas/AnamnesisCase'
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/SourceArtifact'
        ingestions:
          type: array
          items:
            $ref: '#/components/schemas/DocumentIngestionResult'
        bundleImport:
          $ref: '#/components/schemas/FhirBundleImportResult'
    PhysicianPacketCreateResponse:
      type: object
      additionalProperties: false
      required: [case, packet]
      properties:
        case:
          $ref: '#/components/schemas/AnamnesisCase'
        packet:
          $ref: '#/components/schemas/PhysicianPacket'
    PhysicianPacketListResponse:
      type: object
      additionalProperties: false
      required: [physicianPackets, meta]
      properties:
        physicianPackets:
          type: array
          items:
            $ref: '#/components/schemas/PhysicianPacket'
        meta:
          type: object
          additionalProperties: false
          required: [totalPackets]
          properties:
            totalPackets:
              type: integer
              minimum: 0
    ReviewCreateResponse:
      type: object
      additionalProperties: false
      required: [review]
      properties:
        review:
          $ref: '#/components/schemas/ClinicalReviewEntry'
    ReviewListResponse:
      type: object
      additionalProperties: false
      required: [reviews, meta]
      properties:
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/ClinicalReviewEntry'
        meta:
          type: object
          additionalProperties: false
          required: [totalReviews, packetStatus]
          properties:
            totalReviews:
              type: integer
              minimum: 0
            packetStatus:
              $ref: '#/components/schemas/PhysicianPacketStatus'
...
    AuditChainVerificationResponse:
      type: object
      additionalProperties: false
      required: [result]
      properties:
        result:
          $ref: '#/components/schemas/AuditChainVerificationResult'
    OperationsSummaryResponse:
      type: object
      additionalProperties: false
      required: [summary]
      properties:
        summary:
          $ref: '#/components/schemas/OperationsSummary'
...
    DocumentIngestionRequest:
      type: object
      additionalProperties: false
      required: [artifactType, title, contentType, content]
      properties:
        artifactType:
          $ref: '#/components/schemas/ArtifactType'
        title:
          type: string
          minLength: 1
          maxLength: 200
        contentType:
          $ref: '#/components/schemas/DocumentContentType'
        content:
          type: string
          minLength: 1
          maxLength: 12000
        filename:
          type: string
          minLength: 1
          maxLength: 160
        sourceDate:
          $ref: '#/components/schemas/DateOnly'
        provenance:
          type: string
          minLength: 1
          maxLength: 300
...
    FhirImportRequest:
      type: object
      additionalProperties: false
      required: [resource]
      properties:
        artifactType:
          $ref: '#/components/schemas/ArtifactType'
        title:
          type: string
          minLength: 1
          maxLength: 200
        sourceDate:
          $ref: '#/components/schemas/DateOnly'
        provenance:
          type: string
          minLength: 1
          maxLength: 300
        resource:
          $ref: '#/components/schemas/FhirResourcePayload'
    FhirBundleImportRequest:
      type: object
      additionalProperties: false
      required: [resource]
      properties:
        artifactType:
          $ref: '#/components/schemas/ArtifactType'
        sourceDate:
          $ref: '#/components/schemas/DateOnly'
        provenance:
          type: string
          minLength: 1
          maxLength: 300
        allowExternalAttachmentFetch:
          type: boolean
          default: false
        resource:
          $ref: '#/components/schemas/FhirResourcePayload'
    CreatePacketRequest:
      type: object
      additionalProperties: false
      properties:
        requestedBy:
          type: string
          minLength: 1
          maxLength: 120
          description: Optional caller label. When JWT bearer auth is used, if provided it must match the authenticated JWT `sub`; otherwise the server fills it from `sub`.
        focus:
          type: string
          minLength: 1
          maxLength: 300
    SubmitReviewRequest:
      type: object
      additionalProperties: false
      required: [action]
      properties:
        reviewerName:
          type: string
          minLength: 1
          maxLength: 120
          description: Required for API-key or insecure-dev review submission. Optional under JWT bearer auth, where the server binds reviewer identity to the authenticated `sub` and rejects mismatches.
        action:
          $ref: '#/components/schemas/ReviewAction'
        comments:
          type: string
          minLength: 1
          maxLength: 4000
    FinalizePacketRequest:
      type: object
      additionalProperties: false
      properties:
        finalizedBy:
          type: string
          minLength: 1
          maxLength: 120
          description: Required for API-key or insecure-dev finalization. Optional under JWT bearer auth, where the server binds finalizer identity to the authenticated `sub` and rejects mismatches.
        reason:
          type: string
          minLength: 1
          maxLength: 4000